+++ helpers/cubes.py
@@ -1,52 +1,12 @@
-"""Helper module for handling files."""
+"""Helper module for Iris cubes."""
 
-import os
 import warnings
-from pathlib import Path
 
 import iris
-import numpy as np
 from iris.util import equalise_attributes
 from scriptengine.exceptions import ScriptEngineTaskArgumentInvalidError
 
-
-# Using https://stackoverflow.com/revisions/13197763/9
-class ChangeDirectory:
-    """Context manager for changing the current working directory"""
-
-    def __init__(self, new_path):
-        self.new_path = Path(new_path).expanduser()
-        self.saved_path = Path.cwd()
-
-    def __enter__(self):
-        os.chdir(self.new_path)
-
-    def __exit__(self, etype, value, traceback):
-        os.chdir(self.saved_path)
-
-
-def compute_spatial_weights(domain_src, array_shape, grid):
-    "Compute weights for spatial averaging"
-    domain_cfg = iris.load(domain_src)
-    scale_factors = [
-        domain_cfg.extract(f"e1{grid.lower()}")[0][0],
-        domain_cfg.extract(f"e2{grid.lower()}")[0][0],
-    ]
-    cell_areas = scale_factors[0] * scale_factors[1]
-    cell_weights = np.broadcast_to(cell_areas.data, array_shape)
-    return cell_weights
-
-
-def compute_time_weights(monthly_data_cube, cube_shape=None):
-    """Compute weights for the different month lengths"""
-    time_dim = monthly_data_cube.coord("time", dim_coords=True)
-    month_weights = np.array([bound[1] - bound[0] for bound in time_dim.bounds])
-    if cube_shape:
-        weight_shape = np.ones(cube_shape[1:])
-        month_weights = np.array(
-            [time_weight * weight_shape for time_weight in month_weights]
-        )
-    return month_weights
+from helpers.nemo import remove_unique_attributes
 
 
 def load_input_cube(src, varname):
@@ -99,8 +59,13 @@ def set_metadata(cube, title=None, comment=None, **kwargs):
     return cube
 
 
-def remove_unique_attributes(cube):
-    # NEMO attributes unique to each file:
-    cube.attributes.pop("uuid", None)
-    cube.attributes.pop("timeStamp", None)
-    return cube
+def compute_time_weights(monthly_data_cube, cube_shape=None):
+    """Compute weights for the different month lengths"""
+    time_dim = monthly_data_cube.coord("time", dim_coords=True)
+    month_weights = np.array([bound[1] - bound[0] for bound in time_dim.bounds])
+    if cube_shape:
+        weight_shape = np.ones(cube_shape[1:])
+        month_weights = np.array(
+            [time_weight * weight_shape for time_weight in month_weights]
+        )
+    return month_weights
