+++ helpers/files.py
@@ -1,0 +1,19 @@
+"""Helper module for handling files."""
+
+import os
+from pathlib import Path
+
+
+# Using https://stackoverflow.com/revisions/13197763/9
+class ChangeDirectory:
+    """Context manager for changing the current working directory"""
+
+    def __init__(self, new_path):
+        self.new_path = Path(new_path).expanduser()
+        self.saved_path = Path.cwd()
+
+    def __enter__(self):
+        os.chdir(self.new_path)
+
+    def __exit__(self, etype, value, traceback):
+        os.chdir(self.saved_path)
